// web/src/lib/types.ts
export enum UserStatus {
  ACTIVE = "active",
  HOLD = "hold",
  TERMINATED = "terminated",
  INACTIVE = "inactive",
  INVITED = "invited",
  REQUESTED = "requested",
}